countdown and estimation speed text changes

+++ app/src/main/java/cs10/apps/travels/tracer/modules/live/model/EstimationData.kt
@@ -1,3 +1,3 @@
 package cs10.apps.travels.tracer.modules.live.model
 
-data class EstimationData(val totalMinutes: Int, val speed: Int, val fromAverage: Boolean)
+data class EstimationData(val totalMinutes: Int, val speed: Double, val fromAverage: Boolean)
